+++ CHANGELOG.md
@@ -1,5 +1,24 @@
+## [1.1.0] - 2021.01.27
+* add `FbNewsConfig` config
+* **DEPRECATED:** `waiting` use `FbNewsConfig(waiting)`
+* **DEPRECATED:** `noDataOrError` use `FbNewsConfig(noDataOrError)`
+* **DEPRECATED:** `subtitle` use `FbNewsConfig(subtitle)`
+* **DEPRECATED:** `fields` use `FbNewsConfig(fields)`
+* **DEPRECATED:** `borderColor` use `FbNewsConfig(borderColor)`
+* **DEPRECATED:** `backgroundColor` use `FbNewsConfig(backgroundColor)`
+* **DEPRECATED:** `textColor` use `FbNewsConfig(textColor)`
+* **DEPRECATED:** `attachment` use `attachmentsPhotos` or `attachmentsVideos`
+* add `showBorder`
+* add `borderColor`
+* add `backgroundColor`
+* add `textColor`
+* add `FbNewsFields.header`
+* add `FbNewsFields.attachmentsPhotos`
+* add `FbNewsFields.attachmentsVideos`
+* add `FbNewsFields.footer`
+* **fix** render videos with dynamic heigt
 ## [1.0.1] - 2021.01.26
-* fix no likes provided 
+* **fix** no likes provided 
 ## [1.0.0] - 2021.01.17
 * first stabel release
 ## [0.1.0] - 2021.01.16